app/pyproject.toml
# Copyright 2026 Google LLC
#
# Licensed under the Apache License, Version 2.0 (the "License");
# you may not use this file except in compliance with the License.
# You may obtain a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS,
# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
# See the License for the specific language governing permissions and
# limitations under the License.

[project]
name = "ocr-processor"
version = "0.1.0"
description = "Cloud Run Function for OCR Processing"
requires-python = ">=3.11"
dependencies = [
    "functions-framework>=3.5.0",
    "cloudevents>=1.11.0",
    "google-cloud-storage>=2.14.0",
    "google-cloud-documentai>=2.26.0",
    "google-cloud-discoveryengine>=0.11.4",
    "google-api-core>=2.15.0",
    "google-cloud-logging>=3.9.0",
]

[tool.uv]
package = false

[dependency-groups]
dev = [
    "pytest",
    "pytest-mock"
]
